merger refactoring
