+++ hoot-core/src/main/cpp/hoot/core/conflate/merging/LinearMergerAbstract.h
