+++ hoot-core/src/main/cpp/hoot/core/conflate/merging/LinearSnapMerger.cpp
@@ -74,7 +74,7 @@ namespace hoot
 {
 
 // ONLY ENABLE THIS DURING DEBUGGING; We don't want to tie it to debug.maps.write, as it may
-// a very large number of files.
+// produce a very large number of files.
 const bool LinearSnapMerger::WRITE_DETAILED_DEBUG_MAPS = false;
 
 HOOT_FACTORY_REGISTER(Merger, LinearSnapMerger)
